Refactored sql integration tests
